--- be/src/main/java/com/issuetrackermax/domain/oAuth/InMemoryProviderRepository.java
@@ -1,18 +1,0 @@
-package com.issuetrackermax.domain.oAuth;
-
-import com.issuetrackermax.domain.oAuth.service.OauthProvider;
-
-import java.util.HashMap;
-import java.util.Map;
-
-public class InMemoryProviderRepository {
-    private final Map<String, OauthProvider> providers;
-
-    public InMemoryProviderRepository(Map<String, OauthProvider> providers) {
-        this.providers = new HashMap<>(providers);
-    }
-
-    public OauthProvider findByProviderName(String name) {
-        return providers.get(name);
-    }
-}
